🧱 [infra] Removed BVT workflow dependency on the standard build
The build validation step only relies on the dev workflow to complete, but I had it depend on both the dev and standard workflows to simplify the workflow display. This didn't seem like a big deal as the standard build was always faster than the dev build. Until it wasn't....

MacOS builds on GitHub can take a wildly varying amount of time to complete, and I just wanted to standard build take 3x as long to complete as a dev build, even though the dev build is doing a whole lot more. The build verification could have been going while the standard build was puttering a long, but it was blocked due to the artificial workflow dependencies.

It will be blocked no more!